pyproject.toml
[tool.poetry]
name = "sts"
version = "0.1.0"
description = "Project @ DIT"
authors = ["Moritz Enderle <moritz.enderle@th-deg.de>",
           "Florian Eder <florian.eder3@stud.th-deg.de>",
           "Amelie Kammerer <amelie.kammerer@stud.th-deg.de>",
           "QuirinJoshua Groszeibl <quirin.groszeibl@stud.th-deg.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "=3.8.10 || >=3.10.0,<4.0.0" # please us 3.10, this is only because of the AI
pandas = "^2.0.0"
sqlalchemy = "^2.0.9"
toml = "^0.10.2"
pymysql = "^1.0.3"
mysql-connector-python = "^8.0.33"
pillow = "^9.5.0"
pyyaml = "^6.0"
lark = "^1.1.5"
tqdm = "^4.65.0"
fastapi = "^0.96.0"
uvicorn = "^0.22.0"
requests = "^2.31.0"
python-multipart = "^0.0.6"
streamlit = "^1.23.1"
streamlit-extras = "^0.2.7"
streamlit-authenticator = "^0.2.2"
types-requests = "^2.31.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.0"
black = "^23.3.0"
mypy = "^1.2.0"
types-toml = "^0.10.8.6"
sqlalchemy-stubs = "^0.4"
pylint = "^2.17.2"
types-pillow = "^9.5.0.4"
types-tqdm = "^4.65.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
